requirements.txt
torch
torchvision
pyyaml
tqdm
opencv-python
gradio